+++ providers/dns/dyn/dyn_test.go
@@ -1,4 +1,4 @@
-package acme
+package dyn
 
 import (
 	"os"
@@ -31,7 +31,7 @@ func TestLiveDynPresent(t *testing.T) {
 		t.Skip("skipping live test")
 	}
 
-	provider, err := NewDNSProviderDyn(dynCustomerName, dynUserName, dynPassword)
+	provider, err := NewDNSProvider(dynCustomerName, dynUserName, dynPassword)
 	assert.NoError(t, err)
 
 	err = provider.Present(dynDomain, "", "123d==")
@@ -45,7 +45,7 @@ func TestLiveDynCleanUp(t *testing.T) {
 
 	time.Sleep(time.Second * 1)
 
-	provider, err := NewDNSProviderDyn(dynCustomerName, dynUserName, dynPassword)
+	provider, err := NewDNSProvider(dynCustomerName, dynUserName, dynPassword)
 	assert.NoError(t, err)
 
 	err = provider.CleanUp(dynDomain, "", "123d==")
